Move discovery update to mixin class
